pyproject.toml
[project]
name = "car-plate-detector"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "easyocr>=1.7.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "roboflow>=1.1.67",
    "ultralytics>=8.3.159",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"
fix = true

[tool.ruff.lint]
select = [
    "E", # Ошибки PEP8
    "W", # Предупреждения PEP8
    "F", # Ошибки Pyflakes (неиспользуемые импорты, переменные)
    "I", # Сортировка импортов (isort)
    "N", # Стиль именования (PEP8)
    "UP", # Обновление до нового синтаксиса Python
    "B", # flake8-bugbear (ловит потенциальные баги)
    "A", # flake8-builtins (избегаем теневых имен)
    "C4", # flake8-comprehensions (оптимизация list/set comprehensions)
    "S", # flake8-bandit (проверка безопасности кода)
    "PERF", # flake8-perf (проверка на медленные конструкции)
    "PL", # pylint (дополнительные проверки)
]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
strict = true  # Включает строгий режим
warn_return_any = true  # Предупреждать, если функция возвращает `Any`
warn_unused_ignores = true  # Предупреждать о неиспользуемых `# type: ignore`
disallow_untyped_defs = true  # Запрещает функции без аннотаций типов
disallow_untyped_calls = true  # Запрещает вызовы без аннотаций типов
check_untyped_defs = true  # Проверять даже функции без аннотаций
ignore_missing_imports = true  # Игнорировать внешние библиотеки без типизаций
